fix: getServerSideProps and typing
