+++ store/selectors.ts
@@ -1,23 +1,5 @@
-import { PokemonBase } from './pokemonSlice';
+import { AppState } from '../types/types';
 import { createSelector } from '@reduxjs/toolkit';
-
-export interface StatsType {
-  type: {
-    base_stat: number;
-    stat: string;
-  }[];
-}
-interface AppState {
-  pokemonSlice: {
-    allPokemon: PokemonBase[];
-  };
-  pokemonInfoSlice: {
-    image: string;
-    name: string;
-    types: string[];
-    stats: StatsType[];
-  };
-}
 
 // PokemonSlice Selectors
 const getPokemon = (state: AppState) => state.pokemonSlice;
